Move filebeat/input/file methods to libbeat
This refactor makes it possible to reuse code in auditbeat
without depending on filebeat.
