+++ libbeat/common/file/file_other_test.go
