+++ libbeat/common/file/file_windows.go
