+++ libbeat/common/file/file_windows_test.go
